requirements.txt
pandas>=2.0.0
polars>=0.19.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
scipy>=1.11.0
statsmodels>=0.14.0
streamlit>=1.28.0
openai>=1.0.0
duckdb>=0.9.0
plotly>=5.17.0
pytest>=7.4.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pyarrow>=13.0.0
shap>=0.43.0
nbformat>=5.9.0
ipykernel>=6.25.0
openpyxl>=3.1.0
mlflow>=2.8.0
